backend/requirements-render.txt
# PostAssistant Backend Dependencies for Render.com
# Simplified to avoid Rust compilation issues

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0  # Use binary wheel instead of specific version
httpx>=0.25.0
openai>=1.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6
email-validator>=2.1.0
redis>=5.0.0
celery>=5.3.0
stripe>=8.0.0